feat: use tooltip anchor by id
This approach is based on Fast DS tooltip component.

Instead of wrapping the reference element, now we get its ID and use it
to position the floating element (now, the m-tooltip itself).

For that reason, the tooltip styles had to absorb the floating styles, thus,
we no longer import it.