refactor: expose radio
Radio inputs have a particular collective behavior, as they nativelly need
a common `name` property to identify the group and allow the browser to manage
selection.

In this commit, we are implementing radio button custom element by extending the built-in
input radio. For this reason, browsers that do NOT support this feature will have
to load the polyfill references in the radio class.